+++ packages/expo-payments-stripe/android/src/main/java/expo/modules/payments/stripe/util/Fun0.java
@@ -1,5 +1,0 @@
-package expo.modules.payments.stripe.util;
-
-public interface Fun0<R> {
-  R call();
-}